fixed broken test! then broke it again
so i wrote something that fixed the test, but I just wanted to double check something because I had an incling that all i'd done is create a new object, so i wrote another test to check if the 'name' property was still there and it failed. so I'm committing this for review because I think it'll be something decent to discuss.